+++ ui/packages/shared/profile/src/ProfileView.tsx
@@ -1,10 +1,7 @@
 import React, {useEffect, useState} from 'react';
-import {CalcWidth} from '@parca/dynamicsize';
 import {parseParams} from '@parca/functions';
-import {QueryServiceClient, QueryResponse, QueryRequest_ReportType} from '@parca/client';
-import {RpcError} from '@protobuf-ts/runtime-rpc';
-import {Button, Card, Spinner, useGrpcMetadata, useParcaTheme} from '@parca/components';
-import * as parca_query_v1alpha1_query_pb from '@parca/client/src/parca/query/v1alpha1/query_pb';
+import {QueryServiceClient, QueryRequest_ReportType} from '@parca/client';
+import {Button, Card, Input, SearchNodes, useGrpcMetadata, useParcaTheme} from '@parca/components';
 
 import ProfileIcicleGraph from './ProfileIcicleGraph';
 import {ProfileSource} from './ProfileSource';
@@ -127,6 +124,8 @@ export const ProfileView = ({
                     Download pprof
                   </Button>
                 </div>
+
+                <SearchNodes />
               </div>
 
               <div className="flex ml-auto">
@@ -143,7 +142,7 @@ export const ProfileView = ({
 
                 <Button
                   variant={`${currentView === 'table' ? 'primary' : 'neutral'}`}
-                  className="rounded-tr-none rounded-br-none w-auto px-8 whitespace-nowrap text-ellipsis no-outline-on-buttons"
+                  className="items-center rounded-tr-none rounded-br-none w-auto px-8 whitespace-nowrap text-ellipsis no-outline-on-buttons"
                   onClick={() => switchProfileView('table')}
                 >
                   Table
@@ -151,7 +150,7 @@ export const ProfileView = ({
 
                 <Button
                   variant={`${currentView === 'both' ? 'primary' : 'neutral'}`}
-                  className="rounded-tl-none rounded-tr-none rounded-bl-none rounded-br-none border-l-0 border-r-0 w-auto px-8 whitespace-nowrap no-outline-on-buttons no-outline-on-buttons text-ellipsis"
+                  className="items-center rounded-tl-none rounded-tr-none rounded-bl-none rounded-br-none border-l-0 border-r-0 w-auto px-8 whitespace-nowrap no-outline-on-buttons no-outline-on-buttons text-ellipsis"
                   onClick={() => switchProfileView('both')}
                 >
                   Both
@@ -159,7 +158,7 @@ export const ProfileView = ({
 
                 <Button
                   variant={`${currentView === 'icicle' ? 'primary' : 'neutral'}`}
-                  className="rounded-tl-none rounded-bl-none w-auto px-8 whitespace-nowrap text-ellipsis no-outline-on-buttons"
+                  className="items-center rounded-tl-none rounded-bl-none w-auto px-8 whitespace-nowrap text-ellipsis no-outline-on-buttons"
                   onClick={() => switchProfileView('icicle')}
                 >
                   Icicle Graph
@@ -172,14 +171,12 @@ export const ProfileView = ({
                 response !== null &&
                 response.report.oneofKind === 'flamegraph' && (
                   <div className="w-full">
-                    <CalcWidth throttle={300} delay={2000}>
-                      <ProfileIcicleGraph
-                        curPath={curPath}
-                        setNewCurPath={setNewCurPath}
-                        graph={response.report.flamegraph}
-                        sampleUnit={sampleUnit}
-                      />
-                    </CalcWidth>
+                    <ProfileIcicleGraph
+                      curPath={curPath}
+                      setNewCurPath={setNewCurPath}
+                      graph={response.report.flamegraph}
+                      sampleUnit={sampleUnit}
+                    />
                   </div>
                 )}
 
@@ -204,18 +201,14 @@ export const ProfileView = ({
                   </div>
 
                   <div className="w-1/2">
-                    <CalcWidth throttle={300} delay={2000}>
+                    {response !== null && response.report.oneofKind === 'flamegraph' && (
                       <ProfileIcicleGraph
                         curPath={curPath}
                         setNewCurPath={setNewCurPath}
-                        graph={
-                          response?.report.oneofKind === 'flamegraph'
-                            ? response.report.flamegraph
-                            : undefined
-                        }
+                        graph={response.report.flamegraph}
                         sampleUnit={sampleUnit}
                       />
-                    </CalcWidth>
+                    )}
                   </div>
                 </>
               )}