+++ ui/packages/app/web/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.7](https://github.com/parca-dev/parca/compare/ui-v0.13.6...ui-v0.13.7) (2022-06-14)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.13.6](https://github.com/parca-dev/parca/compare/ui-v0.13.5...ui-v0.13.6) (2022-06-14)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.13.5](https://github.com/parca-dev/parca/compare/ui-v0.13.4...ui-v0.13.5) (2022-06-14)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.13.4](https://github.com/parca-dev/parca/compare/ui-v0.13.3...ui-v0.13.4) (2022-06-07)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.13.3](https://github.com/parca-dev/parca/compare/ui-v0.13.2...ui-v0.13.3) (2022-06-07)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.13.2](https://github.com/parca-dev/parca/compare/ui-v0.13.1...ui-v0.13.2) (2022-05-31)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.13.1](https://github.com/parca-dev/parca/compare/ui-v0.13.0...ui-v0.13.1) (2022-05-31)
+
+**Note:** Version bump only for package @parca/web
+
+# [0.13.0](https://github.com/parca-dev/parca/compare/ui-v0.12.38...ui-v0.13.0) (2022-05-30)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.12.38](https://github.com/parca-dev/parca/compare/ui-v0.12.37...ui-v0.12.38) (2022-05-19)
+
+**Note:** Version bump only for package @parca/web
+
 ## [0.12.37](https://github.com/parca-dev/parca/compare/ui-v0.12.36...ui-v0.12.37) (2022-05-16)
 
 **Note:** Version bump only for package @parca/web