+++ ui/packages/shared/client/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.6](https://github.com/parca-dev/parca/compare/ui-v0.13.5...ui-v0.13.6) (2022-06-14)
+
+
+
+## [0.13.4](https://github.com/parca-dev/parca/compare/ui-v0.13.3...ui-v0.13.4) (2022-06-07)
+
+**Note:** Version bump only for package @parca/client
+
+
+
+
+
+# [0.13.0](https://github.com/parca-dev/parca/compare/ui-v0.12.38...ui-v0.13.0) (2022-05-30)
+
+**Note:** Version bump only for package @parca/client
+
+
+
+
+
 ## 0.12.35 (2022-05-05)
 
 **Note:** Version bump only for package @parca/client